pyproject.toml
[project]
name = "poetrypokeadventure"
version = "0.0.1"
description = "Bot discord pokemon"
authors = [
    {name = "Guilherme",email = "guilhermelucario36@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "openai (>=2.5.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "supabase (>=2.22.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
